refactor: 코드 리뷰 반영 및 섹션별로 분리하여 가독성 향상 (#106)
