+++ src/features/chat/hooks/index.ts
@@ -1,1 +1,3 @@
 export { useResizable } from "./useResizable";
+export { useChatMessages } from "./useChatMessages";
+export { useChatPanel } from "./useChatPanel";
